refactor transformation code into packages
