+++ src/main/java/net/coderbot/iris/pipeline/transform/Patch.java
@@ -1,6 +1,6 @@
 package net.coderbot.iris.pipeline.transform;
 
-enum Patch {
+public enum Patch {
 	ATTRIBUTES,
 	VANILLA,
 	SODIUM,
