+++ src/main/java/net/coderbot/iris/pipeline/transform/PatchShaderType.java
@@ -17,16 +17,16 @@ public enum PatchShaderType {
 
 	public static PatchShaderType[] fromGlShaderType(ShaderType glShaderType) {
 		switch (glShaderType) {
-		case VERTEX:
-			return new PatchShaderType[] { VERTEX };
-		case GEOMETRY:
-			return new PatchShaderType[] { GEOMETRY };
-		case COMPUTE:
-			return new PatchShaderType[] { COMPUTE };
-		case FRAGMENT:
-			return new PatchShaderType[] { FRAGMENT, FRAGMENT_CUTOUT };
-		default:
-			throw new IllegalArgumentException("Unknown shader type: " + glShaderType);
+			case VERTEX:
+				return new PatchShaderType[] { VERTEX };
+			case GEOMETRY:
+				return new PatchShaderType[] { GEOMETRY };
+			case COMPUTE:
+				return new PatchShaderType[] { COMPUTE };
+			case FRAGMENT:
+				return new PatchShaderType[] { FRAGMENT, FRAGMENT_CUTOUT };
+			default:
+				throw new IllegalArgumentException("Unknown shader type: " + glShaderType);
 		}
 	}
 }
