+++ src/main/java/net/coderbot/iris/pipeline/transform/TransformPatcher.java
@@ -28,6 +28,18 @@ import net.coderbot.iris.gbuffer_overrides.matching.InputAvailability;
 import net.coderbot.iris.gl.blending.AlphaTest;
 import net.coderbot.iris.pipeline.PatchedShaderPrinter;
 import net.coderbot.iris.pipeline.newshader.ShaderAttributeInputs;
+import net.coderbot.iris.pipeline.transform.parameter.AttributeParameters;
+import net.coderbot.iris.pipeline.transform.parameter.CompositeParameters;
+import net.coderbot.iris.pipeline.transform.parameter.ComputeParameters;
+import net.coderbot.iris.pipeline.transform.parameter.Parameters;
+import net.coderbot.iris.pipeline.transform.parameter.SodiumParameters;
+import net.coderbot.iris.pipeline.transform.parameter.VanillaParameters;
+import net.coderbot.iris.pipeline.transform.transformer.AttributeTransformer;
+import net.coderbot.iris.pipeline.transform.transformer.CommonTransformer;
+import net.coderbot.iris.pipeline.transform.transformer.CompatibilityTransformer;
+import net.coderbot.iris.pipeline.transform.transformer.CompositeTransformer;
+import net.coderbot.iris.pipeline.transform.transformer.SodiumTransformer;
+import net.coderbot.iris.pipeline.transform.transformer.VanillaTransformer;
 
 /**
  * The transform patcher (triforce 2) uses glsl-transformer's ASTTransformer to
@@ -145,7 +157,8 @@ public class TransformPatcher {
 				// parser can be set to the correct version
 				Matcher matcher = versionPattern.matcher(input);
 				if (!matcher.find()) {
-					throw new IllegalArgumentException("No #version directive found in source code! See debugging.md for more information.");
+					throw new IllegalArgumentException(
+							"No #version directive found in source code! See debugging.md for more information.");
 				}
 				Version version = Version.fromNumber(Integer.parseInt(matcher.group(1)));
 				if (version.number >= 200) {
