+++ src/main/java/net/coderbot/iris/pipeline/transform/parameter/ComputeParameters.java
@@ -1,6 +1,7 @@
-package net.coderbot.iris.pipeline.transform;
+package net.coderbot.iris.pipeline.transform.parameter;
 
 import net.coderbot.iris.gl.blending.AlphaTest;
+import net.coderbot.iris.pipeline.transform.Patch;
 
 public class ComputeParameters extends Parameters {
 	public ComputeParameters(Patch patch) {
