+++ src/main/java/net/coderbot/iris/pipeline/transform/transformer/CompatibilityTransformer.java
@@ -1,4 +1,4 @@
-package net.coderbot.iris.pipeline.transform;
+package net.coderbot.iris.pipeline.transform.transformer;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
@@ -10,8 +10,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import io.github.douira.glsl_transformer.ast.node.expression.Expression;
-import net.coderbot.iris.Iris;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -21,6 +19,7 @@ import io.github.douira.glsl_transformer.ast.node.basic.ASTNode;
 import io.github.douira.glsl_transformer.ast.node.declaration.DeclarationMember;
 import io.github.douira.glsl_transformer.ast.node.declaration.FunctionParameter;
 import io.github.douira.glsl_transformer.ast.node.declaration.TypeAndInitDeclaration;
+import io.github.douira.glsl_transformer.ast.node.expression.Expression;
 import io.github.douira.glsl_transformer.ast.node.expression.LiteralExpression;
 import io.github.douira.glsl_transformer.ast.node.expression.ReferenceExpression;
 import io.github.douira.glsl_transformer.ast.node.expression.unary.FunctionCallExpression;
@@ -43,14 +42,17 @@ import io.github.douira.glsl_transformer.ast.transform.ASTInjectionPoint;
 import io.github.douira.glsl_transformer.ast.transform.ASTParser;
 import io.github.douira.glsl_transformer.ast.transform.Template;
 import io.github.douira.glsl_transformer.util.Type;
+import net.coderbot.iris.Iris;
 import net.coderbot.iris.gl.shader.ShaderType;
 import net.coderbot.iris.pipeline.PatchedShaderPrinter;
+import net.coderbot.iris.pipeline.transform.PatchShaderType;
+import net.coderbot.iris.pipeline.transform.parameter.Parameters;
 
 public class CompatibilityTransformer {
 	static Logger LOGGER = LogManager.getLogger(CompatibilityTransformer.class);
 
 	private static final AutoHintedMatcher<Expression> sildursWaterFract = new AutoHintedMatcher<>(
-		"fract(worldpos.y + 0.001)", Matcher.expressionPattern);
+			"fract(worldpos.y + 0.001)", Matcher.expressionPattern);
 
 	private static StorageQualifier getConstQualifier(TypeQualifier qualifier) {
 		if (qualifier == null) {
@@ -71,7 +73,7 @@ public class CompatibilityTransformer {
 		if (parameters.type == PatchShaderType.VERTEX) {
 			if (root.replaceExpressionMatches(t, sildursWaterFract, "fract(worldpos.y + 0.01)")) {
 				Iris.logger.warn("Patched fract(worldpos.y + 0.001) to fract(worldpos.y + 0.01) to fix " +
-					"waving water disconnecting from other water blocks; See https://github.com/IrisShaders/Iris/issues/509");
+						"waving water disconnecting from other water blocks; See https://github.com/IrisShaders/Iris/issues/509");
 			}
 		}
 
@@ -360,10 +362,10 @@ public class CompatibilityTransformer {
 							.getNodeMatch("name*", DeclarationMember.class)
 							.getAncestor(TypeAndInitDeclaration.class)
 							.getMembers()) {
-								String name = member.getName().getName();
-								if (!name.startsWith("gl_")) {
-									outDeclarations.put(name, extractedType);
-								}
+						String name = member.getName().getName();
+						if (!name.startsWith("gl_")) {
+							outDeclarations.put(name, extractedType);
+						}
 					}
 				}
 			}
@@ -403,7 +405,8 @@ public class CompatibilityTransformer {
 								LOGGER.warn(
 										"The in declaration '" + name + "' in the " + currentType.glShaderType.name()
 												+ " shader that has a missing corresponding out declaration in the previous stage "
-												+ prevType.name() + " has a non-numeric type and could not be compatibility-patched. See debugging.md for more information.");
+												+ prevType.name()
+												+ " has a non-numeric type and could not be compatibility-patched. See debugging.md for more information.");
 								continue;
 							}
 							Type inType = inTypeSpecifier.type;
@@ -427,7 +430,8 @@ public class CompatibilityTransformer {
 							LOGGER.warn(
 									"The in declaration '" + name + "' in the " + currentType.glShaderType.name()
 											+ " shader is missing a corresponding out declaration in the previous stage "
-											+ prevType.name() + " and has been compatibility-patched. See debugging.md for more information.");
+											+ prevType.name()
+											+ " and has been compatibility-patched. See debugging.md for more information.");
 						}
 
 						// patch mismatching declaration with a local variable and a cast
@@ -462,7 +466,8 @@ public class CompatibilityTransformer {
 								LOGGER.warn(
 										"The in declaration '" + name + "' in the " + currentType.glShaderType.name()
 												+ " shader has a mismatching dimensionality (scalar/vector/matrix) with the out declaration in the previous stage "
-												+ prevType.name() + " and could not be compatibility-patched. See debugging.md for more information.");
+												+ prevType.name()
+												+ " and could not be compatibility-patched. See debugging.md for more information.");
 								continue;
 							}
